Version 0.1.1 (#3)
* feat(service): add poke-api service

The project uses the pokeApi as api, this commit add the first request

* feat(app): Add pokemon card component

This commit add a card to show a pokemon, with your name, number, image and type

* feat(app): Update pokeApi request

Initially it was possible to get the list of pokemons, but it is
necessary to get the details of the pokemons, this commit adds a new
method in the pokeApi service to get details of the pokemons by name

* feat: Update pokemon-card component

This commit update the layout of pokemon-card

* feat: Add pokemon type card component

This commit adds the card component that represent the pokemon types
with colors according to the type

* feat: Add pokemon-card component

This commit adds the pokemon card with the color according to the main
type of pokemon

* feat: Add HomePage

* feat: Add details page

This commit adds the structure of details page.

* feat: Add api into details page

This commit adds pokeapi to the details page, previously the information
on this page was corrected, now it is dynamic

* feat: Add page routing from home to details

* fix(page): Fix the request on the details page

This commit fix the ability list from request on the details page

* feat(docs): Add README

This commit adds the readme doc into repository

* feat: Add infinite scroll

This commit adds infinite scrolling to the home page, previously the
list of Pokémon here was a maximum of 24, now it is possible to have
them all.

* fix(component): Fix the pokemon-card component

Before, it was possible for the component to not load the list of
Pokemon types to select the background color, crashing the system. Now,
this does not happen because the component already starts with a
pre-selected color.

* feat: Add android platform to the project

* feat: Add android project

This commit generates the android project

* fix(git): Fix git conflicts

This project fixes a conflict in git and a bug caused by GitHub ignoring
the conflict.

Files:
`pokemon-card.component.ts`
`home.page.html`